+++ parseq/help/history.rst
@@ -6,6 +6,8 @@ Version history
 Current GitHub code (modified 28 Feb 2026):
     - Add cumulative logging to time profiling.
 
+    - Fix a bug of triggering transformations after a GUI update from data.
+
 2026.1.0 (5 Jan 2026):
     - Add a pipeline-specific "How-tos" tab to the main tabified dock widget.